moved App.js Todo Component into Todo/index.js to prep for protected routes
